+++ test-third-party-deps/testcases/core_async_example/assert.clj
@@ -1,0 +1,8 @@
+(ns testcases.core-async-example.assert
+  (:require
+   [clojure.core.async :refer [go]]))
+
+(defn sample []
+  (go
+    (assert false)
+    (assert true)))